Change LeNet to prodefined models
