--- deeplodocus/core/project/deep_structure/modules/models/LeNet.py
@@ -1,27 +1,0 @@
-import torch.nn as nn
-import torch
-import torch.nn.functional as F
-
-# LeNet Model definition
-
-
-class LeNet(nn.Module):
-
-    def __init__(self):
-        super(LeNet, self).__init__()
-        self.conv1 = nn.Conv2d(3, 10, kernel_size=5)
-        self.conv2 = nn.Conv2d(10, 20, kernel_size=5)
-        self.conv2_drop = nn.Dropout2d()
-        self.fc1 = nn.Linear(10, 50)
-        self.fc2 = nn.Linear(50, 2)
-
-    def forward(self, x):
-        x = x[0].type(torch.FloatTensor)
-
-        x = F.relu(F.max_pool2d(self.conv1(x), 2))
-        x = F.relu(F.max_pool2d(self.conv2_drop(self.conv2(x)), 2))
-        x = x.view(-1, 10)
-        x = F.relu(self.fc1(x))
-        x = F.dropout(x, training=self.training)
-        x = self.fc2(x)
-        return F.log_softmax(x, dim=1)
